# mentorme_backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Mentor Me API
  description: >
    Backend API for the Mentor Me tutor–student marketplace MVP.
  version: 0.1.0
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:4000
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StudentProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        gradeLevel:
          type: string
        goals:
          type: string
          nullable: true
        preferredSubjects:
          type: array
          items:
            type: string
        notes:
          type: string
          nullable: true
...
    Booking:
      type: object
      properties:
        id:
          type: string
        classId:
          type: string
        studentId:
          type: string
        tutorId:
          type: string
        status:
          type: string
          enum: [PENDING, CONFIRMED, TRIAL, CANCELLED, COMPLETED]
        isTrial:
          type: boolean
        requestedHoursPerWeek:
          type: integer
        startDateExpected:
          type: string
          format: date
        noteFromStudent:
          type: string
          nullable: true
        cancelReason:
          type: string
          nullable: true
        cancelledBy:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /api/students/me:
    get:
      tags: [Student]
      summary: Get current student profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
    patch:
      tags: [Student]
      summary: Update current student profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                gradeLevel:
                  type: string
                goals:
                  type: string
                preferredSubjects:
                  type: array
                  items:
                    type: string
                notes:
                  type: string
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
...
  /api/bookings:
    post:
      tags: [Booking]
      summary: Create booking request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - classId
                - requestedHoursPerWeek
                - startDateExpected
              properties:
                classId:
                  type: string
                isTrial:
                  type: boolean
                requestedHoursPerWeek:
                  type: integer
                startDateExpected:
                  type: string
                  format: date
                noteFromStudent:
                  type: string
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    get:
      tags: [Booking]
      summary: List bookings (role-aware)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: List of bookings
  /api/bookings/{id}:
    get:
      tags: [Booking]
      summary: Get booking detail (owner or admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking detail
  /api/bookings/{id}/confirm:
    patch:
      tags: [Booking]
      summary: Tutor confirms booking
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking confirmed
  /api/bookings/{id}/reject:
    patch:
      tags: [Booking]
      summary: Tutor rejects booking
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Booking rejected
  /api/bookings/{id}/cancel:
    patch:
      tags: [Booking]
      summary: Cancel booking (student/tutor/admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Booking cancelled
  /api/bookings/{id}/complete:
    patch:
      tags: [Booking]
      summary: Mark booking completed
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Completed
...
  /api/admin/bookings:
    get:
      tags: [Admin]
      summary: Admin booking list
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: tutorId
          schema:
            type: string
        - in: query
          name: studentId
          schema:
            type: string
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Filtered bookings
...